This fixes two bugs with scrolling
First: when scrolling down, the viewport would shift down when the cursor got
about halfway down the view (try the bubbletea exmple with 11 items instead of many). The behaviour now scrolls smoothly even with few rows.

Second: with few rows, moving the cursor up from the last item would scroll the the view such that the bottom row would be hidden and the second to last row would take the bottom space.
 Once again, try the bubbletea example with only 11 items.

In fixing these bugs, I got rid of the dependency on the viewport bubble, which was just adding complexity.